Moved ExportData.
